Move GraphLayout out to its own file
